Re-factor misc modules and move all dask utils to `dask_utils` modules
