+++ hyperspy/tests/misc/test_array_tools.py
@@ -23,13 +23,12 @@ import pytest
 
 from hyperspy.misc.array_tools import (
     get_array_memory_size_in_GiB,
-    get_chunk_slice,
-    get_signal_chunk_slice,
     get_value_at_index,
     numba_histogram,
     round_half_away_from_zero,
     round_half_towards_zero,
 )
+from hyperspy.misc.dask_utils import get_chunk_slice, get_signal_chunk_slice
 
 dt = [("x", np.uint8), ("y", np.uint16), ("text", (bytes, 6))]
