+++ hyperspy/tests/signals/test_lazy.py
@@ -24,10 +24,10 @@ from dask.threaded import get
 import hyperspy.api as hs
 from hyperspy import _lazy_signals
 from hyperspy._signals.lazy import (
-    _get_navigation_dimension_chunk_slice,
     _reshuffle_mixed_blocks,
     to_array,
 )
+from hyperspy.misc.dask_utils import _get_navigation_dimension_chunk_slice
 
 
 def _signal():
